getting all changes

+++ Components/TaskCard.jsx
@@ -9,7 +9,7 @@ import { useNavigation } from "@react-navigation/native";
 import { getDatabase } from "../database";
 import { deleteCreatedTask } from "../queries";
 
-const TaskCard = ({ data }) => {
+const TaskCard = ({ data, type }) => {
   const [topic, setTopic] = useAtom(tasks);
   const [completed, setCompleted] = useState(data.status === "complete");
   const [id, setId] = useAtom(taskId);
@@ -20,20 +20,27 @@ const TaskCard = ({ data }) => {
 
   const handleComplete = async () => {
     if (db) {
-      const result = await db.getAllAsync(
-        `SELECT * FROM ripplestatus where expiry=? and taskId=?`,
-        [todayDate, data.taskId]
-      );
-      if (result.length > 0) {
-        await db.runAsync(
-          `DELETE FROM ripplestatus WHERE expiry=? and taskId=?`,
+      if (type === "daily") {
+        const result = await db.getAllAsync(
+          `SELECT * FROM ripplestatus where expiry=? and taskId=?`,
           [todayDate, data.taskId]
         );
+        if (result.length > 0) {
+          await db.runAsync(
+            `DELETE FROM ripplestatus WHERE expiry=? and taskId=?`,
+            [todayDate, data.taskId]
+          );
+        } else {
+          await db.runAsync(
+            `INSERT INTO ripplestatus (taskId, status, expiry) VALUES(?, ?, ?)`,
+            [data.taskId, "complete", todayDate]
+          );
+        }
       } else {
-        await db.runAsync(
-          `INSERT INTO ripplestatus (taskId, status, expiry) VALUES(?, ?, ?)`,
-          [data.taskId, "complete", todayDate]
-        );
+        db.runAsync(`UPDATE ONETIME SET STATUS=? WHERE TASKID=?`, [
+          data.status === "complete" ? "incomplete" : "complete",
+          data.taskId,
+        ]);
       }
     }
   };
@@ -56,7 +63,17 @@ const TaskCard = ({ data }) => {
 
   const handleDelete = async () => {
     if (db) {
-      await db.runAsync(deleteCreatedTask, data.taskId);
+      if (type === "daily") {
+        await db.runAsync(
+          `update ripplereminder set isdeleted=1 where taskId=?`,
+          data.taskId
+        );
+      } else {
+        await db.runAsync(
+          `update onetime set isdeleted=1 where taskId=?`,
+          data.taskId
+        );
+      }
     }
   };
 
@@ -74,7 +91,7 @@ const TaskCard = ({ data }) => {
     <View style={tw` flex justify-center items-center w-full px-2`}>
       <TouchableOpacity
         style={[
-          tw`flex flex-row justify-between items-center py-3 w-full rounded-xl shadow-xl my-2 px-2`,
+          tw`flex flex-row justify-between items-center py-2 w-full rounded-md shadow-xl my-1 px-2`,
           {
             backgroundColor:
               data.status === "complete" ? colors.green : colors.blue,
@@ -89,7 +106,7 @@ const TaskCard = ({ data }) => {
       >
         <View
           style={[
-            tw`flex flex-row justify-center items-center`,
+            tw`flex flex-row justify-start pl-2 items-center`,
             { width: "100%" },
           ]}
         >
@@ -107,10 +124,10 @@ const TaskCard = ({ data }) => {
             iconRight={false}
           /> */}
           {/* <View style={[tw`flex`, { gap: 5 }]}> */}
-          <Text style={[tw`text-2xl font-bold text-white`]}>
-            {data.taskHeading.length < 18
+          <Text style={[tw`text-base text-white pb-2`, { fontWeight: "500" }]}>
+            {data.taskHeading.length < 30
               ? data.taskHeading
-              : `${data.taskHeading.substring(0, 18)}...`}
+              : `${data.taskHeading.substring(0, 30)}...`}
           </Text>
           {/* <View style={[tw`flex flex-row items-center `]}>
               <Icon
